Junit Cleanup

+++ src/main/webapp/resources/js/directives/confirmDeleteReportDialog.js
@@ -11,40 +11,41 @@ app.directive("confirmDeleteReport",function(){
 		},
 		link: function($scope, element, $watch, $http){
 			$scope.$watch("report", function(){
-				jQuery("#confirmDeleteReportDialog").dialog({
-					autoOpen: false,
-					title: "Delete Report?",
-					modal: true,
-					buttons:[
-						{
-							text: ("Delete Report #" + $scope.report.id),
-							icons: {primary: "ui-icon-closethick"},
-							click: function() {					
-																
-								var callback = function(){
-									jQuery("#confirmDeleteReportDialog").dialog("close");									
-									$scope.updateFunc();
+				if ($scope.report){
+					jQuery("#confirmDeleteReportDialog").dialog({
+						autoOpen: false,
+						title: "Delete Report?",
+						modal: true,
+						buttons:[
+							{
+								text: ("Delete Report #" + $scope.report.id),
+								icons: {primary: "ui-icon-closethick"},
+								click: function() {					
+																	
+									var callback = function(){
+										jQuery("#confirmDeleteReportDialog").dialog("close");									
+										$scope.updateFunc();
+									}
+									var url = "rest/report";								
+									jQuery.ajax ({
+										    url: url + "/" + $scope.report.id,
+										    type: "DELETE",
+										    dataType: "json",
+										    contentType: "application/json; charset=utf-8",
+										    success: callback
+										    }
+									);
 								}
-								var url = "rest/report";								
-								jQuery.ajax ({
-									    url: url + "/" + $scope.report.id,
-									    type: "DELETE",
-									    data: submitText,
-									    dataType: "json",
-									    contentType: "application/json; charset=utf-8",
-									    success: callback
-									    }
-								);
+							},
+							{
+								text: ("Don't Delete the Report"),
+								click: function(){
+									jQuery("#confirmDeleteReportDialog").dialog("close");
+								}
 							}
-						},
-						{
-							text: ("Don't Delete the Report"),
-							click: function(){
-								jQuery("#confirmDeleteReportDialog").dialog("close");
-							}
-						}
-					]
-				}); //end dialog
+						]
+					}); //end dialog
+				}//end if
 			}); //end watch	
 		}//end link
 	};
